Update patch set 8
Patch Set 8: Code-Review+1

(1 comment)

Thanks Tim for the answers, thanks Almothana for the changes!

Almothana, I'd suggest to add a TODO to the checkBreakpointDidNotActivate();
to mention that we need to implement it properly. We definitely want to
check that we didn't pause.
Since we need to investigate more, we can do that in a later CL.

Patch-set: 8
Reviewer: Gerrit User 1396836 <1396836@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1