Re-enable the network e2e test
This e2e test was disabled because it was flaky.
One of the reasons it failed was because of a buggy helper method
`selectRequestByName` which, instead of comparing request names with
some value, was assigning the value to the first request (= instead of
===).
The other reason was likely because the test took longer than the set
timeout for e2e tests.

Bug: 1093287
Change-Id: Ice729c295f7f1ff22c40e76476893d258fd0284f