[wasm] Update to wasmparser 2.2.0
Don't print all the types in the beginning of the Wasm disassembly,
and use the abbreviated form for imports and exports to not bury the
lede unnecessarily. This makes debugging huge Wasm modules a bit more
pleasant experience. It also improves the performance quite a bit for
larger modules, since the types contribute a significant amount of
non-breakable lines in the common case.

Fixed: 1092763
Refs: wasdk/wasmparser#56
Bug: 1071432, 1058836, 1092460
Change-Id: Ibe32a355cff811b892d8875e8fb2d85d83f5f836