Update patch set 10
Patch Set 10:

(1 comment)

Thanks for the follow ups Alex and Patrick!

After talking with Patrick offline, we are planning on checking in as is and following up on the remaining comments for tests and cleanup in future CLs. That will allow us to more quickly iterate/gather feedback behind the experiment once we land the backend and add settings to enable this feature.

Patch-set: 10