Register Forwarded Shortcuts in ShortcutRegistry
This CL fixes a bug where forwarded shortcuts (which can be used in the
target page in addition to within the DevTools) aren't whitelisted when
the active set of keyboard shortcuts is changed. With this fix, the
browser key whitelist is updated every time the active shortcut preset
is changed.

Bug: 174309
Change-Id: I226b4ddfdae6d5e07b3983dd2163dde8979aec74