Fix overlay tooltip arrow spacing
To support forced-colors mode, a 1px border was added to the overlay's
tooltip content box. This had the unintended consequence of adding a
small bit of whitespace between the tooltip content box and its arrow
when the arrow points down.

This patch removes 1px of top positioning from the tooltip's arrow when
the arrow points down.

Bug: 1087783
Change-Id: Icd40feed52b0ca3c5ce0ab6b04d2be8c3f3f81dc