Update patch set 6
Patch Set 6:

For the record, I don't think downstream requirements are good reasons to block changes and fixes in upstream. We have no insight into what these requirements are, and we should not feel responsible to carry the cost of any divergence.

The best way to avoid these breakages is to align downstream with upstream as much as possible, e.g. by upstreaming.

Patch-set: 6
CC: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>