Introduction of the Network Console experiment.
Network Console is new OSS being built to configure and make
synthetic network requests. The bulk of the UI runs in a separate
context from the DevTools, hosted in an iframe and served over the
DevTools remote devtools:// scheme.

The Network Console OSS is sourced from
https://github.com/microsoft/edge-devtools-network-console

DISABLE_THIRD_PARTY_CHECK=This CL adds a new third party OSS and
components to consume it.

Bug: 1093687
Change-Id: I91b261603f2c6d421016ebd07886d5f805ad9935