[DevTools] Fix an issue in emulated device dimension in spec.
Previously, the dimension numbers in emulated_devices for Surface Duo
were calculated based on real dimension and device-pixel-ratio of 3. But
the device-pixel-ratio was 2.5, which is consistent with Duo spec.

As device-pixel-ratio is used as device_scale_factor in the emulation,
this inconsistency will cause issue for platform. Fix the dimension
numbers using device-pixel-ratio/device_scale_factor of 2.5.

Change-Id: I82ff75d781a01ffaaf3a590a4b2ab9137accd44f