Center align timeline status dialog button
This patch center aligns the stop button in the timeline status dialog
so it remains visible when the DevTools window is thin.

Unfortunately we can't use the `width` media feature to solve this since
when DevTools is docked, the `width` measures the entire width of the
tab's window.

Before: https://imgur.com/0onCrEv
After: https://imgur.com/w6BZc88
Bug: 983948
Change-Id: I5ce20a881aa431c146424afe7eca23ec7920d95a