Update patch set 14
Patch Set 14:

OK. Seeing this with the Security tab works pretty well, except for one major item related to the other experiment.

When I move the Security tab to the Drawer, it appears in the Drawer's More Tools menu. However, when it's not in the Drawer, it's only accessible via the primary More Tools menu.

I think this is a case for unifying the two More Tools menus to present the same content.

Patch-set: 14