Update patch set 15
Patch Set 15:

(5 comments)

> Patch Set 14:
> 
> OK. Seeing this with the Security tab works pretty well, except for one major item related to the other experiment.
> 
> When I move the Security tab to the Drawer, it appears in the Drawer's More Tools menu. However, when it's not in the Drawer, it's only accessible via the primary More Tools menu.
> 
> I think this is a case for unifying the two More Tools menus to present the same content.

You have a good point about the implementation of the ... menu of the drawer, I agree with you, it should mirror the content of the other ... More tools menu. 
We can address that in a separate CL to separate the concerns and make it easier to revert if needed.

Patch-set: 15