Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > IIUC, there's no need for the "package" subdirectory. Why not import it directly into front_end/third_party/acorn-numeric-separator? For reference, https://github.com/acornjs/acorn-numeric-separator doesn't have a package directory.
> 
> The package subdirectory comes from `npm pack`, and when we pull them as tgz from npmjs.com (which isn't possible until upstream merges the fix).

Is there not a model to do direct github source import? This is what I'm used to in the mono-repo.

Also, you're clearly not getting the source from npmjs.com, so I'm not sure why that matters here -> will follow-up CLs replace this whole import with the official github source?

Patch-set: 3