Fix scrollbar issue in sources panel
The previous fix (https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2241753)
was incorrect. It overrode a style rule in `codemirror.css`, hiding
all scrollbars. Instead, the culprit was that we do not load the
`codemirror.css` file from `third_party` and this file was updated in
https://github.com/codemirror/CodeMirror/commit/632f30be88392aa8e86a5e28697df78f759cc913

Therefore, we have to update our `cm/codemirror.css` to update the
correct CSS rules. This ensures that the console tab has no scrollbar,
but the sources panel does.

We have to figure out how to properly use the `codemirror.css` file
from `third_party/`, which is tracked in crbug.com/1076825

R=jacktfranklin@chromium.org
CC=bmeurer@chromium.org

Fixed: 1094168
Bug: 1093786
Change-Id: Ide22274ea35fd47c4198442f7b59191dfb8ac599