Update patch set 15
Patch Set 15:

(6 comments)

I found the issue with the "More tools" as well, but we can fix that in a follow-up. I left a comment about the combination of movable and closeable tabs. I wasn't able to crash it (yay!), but I don't understand why that check is necessary if it doesn't seem to have an effect.

Patch-set: 15