Add ElementsBreadcrumbs component with examples
This CL adds the `ElementsBreadcrumbs` component and documentation to
the component server.

Important: this CL *does not add the component to the UI*. That will be
done in a follow up CL once this has shipped and we have done some other
infrastructure work that is currently blocking components going into
production.

This is also why the file is called `NewElementsBreadcrumbs`, to avoid a
nameclash with `ElementsBreadcrumbs.js`. In the CL that puts this
component into production I will change the filename as the clashing
won't be an issue.

Change-Id: I8039d314ba02c834c3f8ec26f9bdb547858aa2dd