Update patch set 4
Patch Set 4:

> Patch Set 1: Code-Review-1
> 
> What are the changes wrt. my WIP CL? This change does not use the DevTools Experiment setting to hide the extension API. A few things are also not in sync with the design doc, I think. The repr and LocalJSONObject translation shouldn't contain hardcoded types. The handleScript function also shouldn't call out to the extension.

We'll discuss on Monday. I've verified that the API is hidden this way unless the extension has the "experimental" permission and --enable-experimental-extension-apis is passed on the command line. The changes I've made are to relocate the API entry point, split the API changes that need to go into Chrome from the extension changes, and changes to get it to pass the type checker, which turned out to be simplified by removing the old CXXDwarf plugin. Maybe we don't want to do that yet? Looks like I need to disable the Dwarf Language Plugin tests too if we're doing that.

Patch-set: 4
Label: Commit-Queue=0