Enable the no-return-assign ESLint rule
This rule disallows assigning a value in a return statement.
It also works in one-liner arrow functions so that a => a = 2 triggers a
warning but a => (a = 2) doesn't.
This is to avoid cases where you actually meant to compare to 2, not
assign the value 2, but made a mistake. These sorts of issues can easily
pass review and cause weird problems that are hard to track down.

The rule is enabled in this CL, and all warnings are fixed.

More info about the rule: https://eslint.org/docs/rules/no-return-assign

Bug: 1093379
Change-Id: I38b98ea59ecaedc83a7f69073b17980808f9a3bf