Add a live-updating timeline view to the Media component in devtools
I had to add a small helper to the FlameChart.js file to be able to add
a scroll wheel listener for zooming in and out.

Right now I've got the play/pause/ended/destroyed events hooked up,
as well as the audio/video buffering delays. I want to next hook up
some of the new work the team has been doing on smoothness and cadence
of frame presentation.

Example usage:
https://files.tedm.io/dtss/timeline/example_usage1.mp4

Example with throttled internet:
https://files.tedm.io/dtss/timeline/example_usage2.mp4

Change-Id: I515acb76196d127059a958388168f0794e190274