[debugger] Unify possible breakpoint queries.
In `_createNewBreakpoint` we've queried for all possible breakpoints in
the same line, up to column offset 1024, while in `_addBreakpoint` we
ask for all possible breakpoints in this line by passing the next line
number. The latter makes more sense and will also work properly for Wasm
disassembly, where each line corresponds to a single instruction, so we
adopt the latter for `_createNewBreakpoint` as well.

Fixed: chromium:1091735
Bug: chromium:1071432
Change-Id: I6773129082cd7a2685b0f3770d66bbcd4686767d