Update patch set 7
Patch Set 7:

(7 comments)

Sorry for the review delay. There was a public holdiay in Germany and most ppl where OOO for a few days.

Thanks for the re-write using the TS parser. I think there are might be some subtle bugs in the re-written code and I left some comments :)

Patch-set: 7