Update patch set 2
Patch Set 2:

Interesting, this is already the second CL I see related to a super narrow DevTools viewports.

Probably a better solution is to not center align the button but to make the "Profiling ..." box smaller. Buttons should be right aligned in our design system. I can create you a mock for this tomorrow.

Regarding mind-width of DevTools: I was also thinking about this but I believe giving our pro users the freedom of resizing DevTools however they like is more desirable than restricting this. In the end, we don't need to fix every bug that says "DevTools not usable if 20px wide".

Will make a mock for you tomorrow.

Patch-set: 2