[wasm] Ensure we always add the proper truncation marker.
The `WasmDisassembler#maxLines` approach that we have introduced
originally doesn't work well with the chunked disassembly mechanism
(which we currently don't use) and also doesn't work correctly with
the machinery in wasmparser that avoids breaking def-use chains for
labels.

So for now, we'll just generate the full disassembly and throw away
the lines that go above the magic 1M limit (for CodeMirror).

Fixed: chromium:1094165
Bug: chromium:1058836, chromium:1071432
Change-Id: I02021573d72fff0854fe7e42096dd3c48fa14d0e