[sources] Don't show pretty-print infobar for Wasm disassembly.
Content for .wasm files is fetched as Base64-encoded string and served
like that all the way up to the UI layer. Thus the minified script de-
tection logic often thinks that this is a minified JavaScript file and
would offer to format it, which doesn't make any sense, and also doesn't
really work anyways. This changes the predicate to not even consider
Wasm wire bytes for formatting all together.

Fixed: 1094880
Bug: 1071432
Change-Id: I6bda7969eea1ddf8039b22f9c509ed922fc5843c