Avoid overlaping grid column and row labels
This change fixes the is[First/Last]Column and is[Top/Bottom]Row logic,
by making sure those booleans take implicit lines into account.

Then the is[Top/Bottom]Row booleans are used to shift down or up the
first and last row labels, to avoid them from overlaping with the
column labels.

For now, this CL only addresses positioning of labels when they are
drawn outside the grid container. A follow-up will be done to also
fix positionng when labels are places inside the grid.

This change also adds a new test grid to the debug HTML file, and
makes that file easier to maintain.

Bug: 1047356
Change-Id: Ibedb44d51b442ddc3cacadf28636e0eb8430091e