Update patch set 2
Patch Set 2: -Code-Review

This CL is now ready for review. I have confirmed that the build times in Chromium for a single edit remain stable before and after. The essential difference between the original change and this one, is that we specify the `--external` sources on the command line, but keep the rest in the `rollup.config.js`.

Patch-set: 2
Label: Code-Review=0