Update patch set 2
Patch Set 2:

FYI Benedikt and Simon, we are going to reland this change soon. We have verified there is no noticeable increase in build time. If you do notice a significant increase after this lands, please revert this CL again and we have to take another look.

Patch-set: 2
CC: Gerrit User 1183894 <1183894@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1323977 <1323977@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0