Add browser test for forwarded key whitelist changes
This CL adds a browser test for the changes in [1], which added logic to
update the whitelist of keys that the browser forwards to the DevTools
whenever the keybind set changes rather than just once at launch.
Forwarded keys don't work in remote debugging/hosted sessions, so they
can't be tested with an e2e test. There's also a small change here to
enable this new test in the backend: https://chromium-review.googlesource.com/c/chromium/src/+/2246552

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2231908

Bug: 174309
Change-Id: Ib511abe8192fce245e72b371954eb94c2aa3d907