Update patch set 4
Patch Set 4:

Yes this bundle should be created as part of a Ninja target. We should not be including build output in our sources folder anymore (we are working as we speak on making sure all pieces of infrastructure are able to cope with that).

You can define a GN action (https://gn.googlesource.com/gn/+/master/docs/reference.md#func_action) that invokes `//third_party/node/node.py` to run your build script. I was working on fixing that, but it broke Chromium. The original CL was https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2208134 I will work on relanding that.

If you are uncertain how to integrate with Ninja, I would be happy to take a look. In that case, please check in your build script, but not the output bundle. Then I can write a GN action to generate the bundle in the appropriate location and send you the CL. Let me know what works best for you.

Patch-set: 4