Update patch set 1
Patch Set 1: Commit-Queue+1

Have to figure out what the impact on the bundle size is, so this is not ready to submit as-is. But at least it contains the changes required to move the bundle over to non-remote.

I also still need to build Chromium with this change and verify it correctly loads.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1
Work-in-progress: false