Update patch set 2
Patch Set 2:

Adding Yang with regards to the increase in the bundle size. This would increase the bundle size by about ~400KB, but no longer would rely on the appspot server. Do you know if there are any requirements for us to be below a certain file size limit? We could in theory submit this CL as-is and if the Chromium binary size sheriffs or Chrome OS folks are unhappy with it, we could revert it. Let me know what your preferred way forward is.

Patch-set: 2
Reviewer: Gerrit User 1118373 <1118373@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>