Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Probably not relevant, but I ask anyways: Isn't the first row/column number in the wrong column? Shouldn't it be in the top left corner?

Thanks for asking! This is quite an interesting thing that happens with the overlay: the particular grid I'm highlighting here has an implicit first row and an implicit first column. Those were created because an item got positioned with negative line numbers that exceeded the explicit size of the grid.

Patch-set: 1