Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> (1 comment)
> 
> I'm not sure if the implementation has to work with the models of all targets or if the mainTarget
> would be enough. 
> @Nina: Do you know how this is implemented in backend? Does it distinguish the targets?

Well I wrote the back-end but don't know the difference between targets. The environment requires devtools to be attached to a frame and will affect that frame and all its children [1]. We don't support this API on android so remote debugging won't work. Does that help?

[1] https://source.chromium.org/chromium/chromium/src/+/master:content/browser/devtools/protocol/webauthn_handler.cc;l=106-116;drc=c7677ed47539a7c7e433cc7d5d5e963c9faf58ce

Patch-set: 4