Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
> 
> Is there a reason why we don't put the i18n-bundle.js in the out/ directory and invoke the {buildi18nBundle.js} from a BUILD.gn file?

Hi Simon, thanks for the feedback. This was a difficult choice 😊, I considered  both options but ended up doing the "pre-bundle" as I believe it also provides some benefits  on a library that will not change often (faster, less error prone), I was not aware that devtools is trying to get rid of bundles in the src folder, so given this, creating the bundle on build makes sense. I'll make the changes.

Patch-set: 4
Label: Commit-Queue=0