[Autofill] Remove unnecessary navigation in a test in Test.js
This CL removes unnecessary in testDispatchKeyEventShowsAutoFill test.

Why it has be working before:
The reason this has been working so far is that the autofill
 menu contains exactly one element.

Why it cannot stay as is:
However, upcoming features will introduce another entry in the
autofill menu to hide the suggestions. This extra navigation causes
the extra menu entry to be selected instead of the top entry.

Why it still works now:
Because Chrome select the first item by default when the autofill
menu is shown.

Bug: 1030649
Change-Id: I74eab8c2be4c43b0937d0ee772fedbf2b359fa6f