Update patch set 20
Patch Set 20:

(1 comment)

We added a step that checks that the entire source code have been displayed instead of waiting for the state of the target page.
Please note that there seems to be an issue where if the user have added a breakpoint and the script has paused on it, removing the breakpoint and refreshing keeps the script paused on the same location. Is this intended behaviour? if it is please let me know so we can add a step that continues the script before proceeding to the next steps.

Patch-set: 20