Update patch set 6
Patch Set 6:

(4 comments)

We are inheriting some build infrastructure from upstream, primarily related to CommonJS modules. I would prefer to not do that, if we are in a position to do so. Could anyone give me an explanation what we could maybe do here?

I would be okay with landing as an interim solution, but before shipping to stable I would prefer to make sure we maintain a consistent module-system in DevTools. That allows us to apply a consistent build time and optimization path for all source code that ships in the DevTools bundle.

Patch-set: 6