Update patch set 3
Patch Set 3: -Auto-Submit

> Patch Set 3:
> 
> Hmm I think it still looks a bit too much at the top. I did 
> 
> ```
> .adorner-container devtools-adorner {
>     margin-left: 3px;
>     - vertical-align: middle
>     + line-height: 1.5;
> }
> ```
> 
> and it looks visually more centered to me:
> 
> Before: https://imgur.com/a/9Okbcyf
> After: https://imgur.com/a/QZ3rgn9
> 
> Maybe a Mac issue? Could you check on Linux with those changes?

Let me try it on Linux. In some cases it looks center when at 100% scale, but looks to the top or to the bottom when at 150% scale (I make it 150% scale to test out how it would look on a screen with less resolution). And some adorners look center while some others look more to the top. I'm still trying to figure out why.

Patch-set: 3
Label: Auto-Submit=0