[Performance] Retime markers against nav start records
The current timeline behavior assumes that time 0 is the start of the
trace. This assumption holds up just fine under many cases, but in page
load contexts in particular, it does not. Indeed there is a difference
between the values for metrics like FCP and LCP as reported in the trace
as compared to the values reported by PerformanceObserver, since the
former is measured from the trace start, where the latter measures since
the NavigationStart trace event.

This CL changes two things: 1) it changes the times shown in the ruler
to be relative to NavigationStart events, should any exist; 2) it
modifies the time reported in the details for markers to again be
relative to the nearerst NavigationStart event. Every time there is a
navigation event, the ruler's time is reset to 0ms, thus we catch traces
with multiple navigation events.

See https://imgur.com/a/ueU7GbD for before & after images.

Bug: 974550
Change-Id: Ic75ac3940ec18ba3413745ceb7c7ae528f5e9019