Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hm... I am slightly surprised. On https://rsolomakhin.github.io/autofill/, I can focus the first Name field but nothing is autoselected?!?

You probably used the mouse to focus the field, right?

The method "signalToShowAutofill()" doesn't focus the field, but invokes the button "Arrow Down"
To try it, use tab navigation to focus the field, and then press on the "Arrow down" button.
The menu will be shown, first item will be selected.

Patch-set: 5