Fix stylelint path resolution on Windows
On Windows, if your checkout is at:

    C:\devtools\devtools-frontend

Then the stylelint CLI lives here:

    C:\devtools\devtools-frontend\node_modules\stylelint\bin\stylelint.js

Prior to this patch, the lint CSS script was instead trying to use:

    C:\node_modules\stylelint\bin\stylelint.js

That is, it went two levels too far up the directory tree. This only
happened on Windows.

This patch avoids the problem altogether by forcing the root directory
to be the current working directory and passing a root-relative glob
instead.

As a drive-by, this patch cleans up the `_getFilesToLint` helper in
`PRESUBMIT.py`.

Bug: chromium:1083142
Change-Id: Ida7d19e4c1cb8954346a3103ce95748f0bc38f00