Update patch set 2
Patch Set 2:

I don't have the full picture, but I agree with Philip that it may be confusing to
have a type Variable that has nested variables.

Would it make sense to have a type namespace, that can have an array of namespaces *and* an array
of variables as properties? I don't know how everything ties together, but looking at the api.pdl this seems more intuitive to me.

Patch-set: 2