Update patch set 2
Patch Set 2:

It might even be simpler to send the namespace nesting of a variable as an array of strings. For a variable foo::bar::baz that's ["foo", "bar", "baz"]. I'm including the variable name so that we can show the short name in the frontend while usin the FQ name for identifying the appropriate evaluator.

Patch-set: 2