Update patch set 8
Patch Set 8:

Added a new iteration, this time it is bundling using rollup instead of browserify. I tested the library locally (removing the previous bundle and pluging the new one) and everything works as expected.


tvanderlippe@chromium.org thanks for the great suggestions, there were a couple of comments that were not clear to me would you mind ellaborating?

-

Patch-set: 8