Rework V8 code cache-related UI
This patch causes code caching information to be displayed regardless
of whether or not caching was successful. Previously, DevTools wouldn’t
show anything related to code caching if code caching didn’t happen.

It also improves the UX by presenting the information in a more logical
order and by formatting the cache sizes nicely.

To distinguish from the many other caches DevTools exposes, the name
“compilation cache” was chosen. One downside of the term “compilation
cache” is that it’s used to refer to the “isolate cache” in V8. On the
other hand, the term “bytecode cache” is not as future-friendly, since
V8 might one day have cacheable NCI code.

Screenshots:

- Before: https://i.imgur.com/dDHOiMj.png
- After:
    — Script not eligible: https://i.imgur.com/KyXdZ1N.png
    - Script loaded from cache: https://i.imgur.com/npr4H1B.png

Test/demo page:
https://mathiasbynens.github.io/css-dbg-stories/v8-compilation-cache.html

Bug: chromium:912581
Change-Id: I053fcf2a731757105c00004896b9bd0ef05f3a34