Update patch set 9
Patch Set 9:

(2 comments)

> Patch Set 8:
> 
> > Patch Set 6:
> > 
> > (4 comments)
> > 
> > We are inheriting some build infrastructure from upstream, primarily related to CommonJS modules. I would prefer to not do that, if we are in a position to do so. Could anyone give me an explanation what we could maybe do here?
> > 
> > I would be okay with landing as an interim solution, but before shipping to stable I would prefer to make sure we maintain a consistent module-system in DevTools. That allows us to apply a consistent build time and optimization path for all source code that ships in the DevTools bundle.
> 
> @timvanderlippe Would you mind ellaborating on this? I'm not following which infrastucture this refers to. When you mean landing this do you refer to the current iteration (browserify) or the suggested implementation (rollup)?

Thanks for using Rollup instead! I was alluding to the fact that this CL was using CommonJS modules + browserify previously. Since we are no longer using them, my comments are resolved 😊

Sorry for going back and forth on this CL, but we are getting quite close! I have no concerns with the implementation, only with regards to the location of the third_party packages and the integration with TypeScript. Once these are resolved, I think this is ready to go.

Patch-set: 9