Grid: shift grid labels a bit when they're inside the grid and at the corner
Labels are flipped inside the grid when they are too close the edge of the
viewport. When this happens, and when there are both column and row labels at a
corner of the grid, they overlap.
This change avoids this from happening by shifting their position slightly.

This change contains the following things:
- An update of the grid debug page to test this out.
- An update of the grid debug page to reflect recent backend changes whereby
grid line paths don't get closed between cells anymore.
- A fix to the frontend path builder logic that was closing paths an extra time
and causing lines to be drawn on top of eachother.
- The implementation of the label shifting logic.
- A refactor of the code in css_grid_label_helper to make it easier to read.
- Some new unit test cases.

The shifting of labels happens with CSS transforms. When the JS code detects
that a label is at a corner, inside the grid, and that another label is also
present there, it applies a CSS class to the label.
The CSS then takes over and, depending on the type of label, translates it
by a few pixels in the right direction.

Before: https://imgur.com/Yi8TnY8.jpg
After: https://imgur.com/gjHJ71x.jpg

Bug: 1096629
Change-Id: I161c08ae4641269ff9d84ab068fb61ce83f561f5