Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I noticed this CL is a bit old, are you still trying to land it?

It fell off my radar for a bit as I was busy with other things. But ideally, yes, I'd like to land it. I think I did address your first review comments, and apart from a rebase, it should be good for a new round of reviews.

Patch-set: 4