[Perfmon] Improve contrast of labels in dark theme
The logic for tracking colors for the dark theme did not test the labels
for the Perfmon charting, but rather the chart lines themselves. This
CL ensures the labels are checked, but also bumps the minimum cap value
so that the labels in the Perfmon obtain a contrast of > 7.0.

See

Bug: 1098637
Change-Id: I6b366d672fca971b134261fa1de2477ef52345c3