Grid: Skip line labels to avoid overlap
When a grid container has narrow tracks, the labels draw in the overlay may
overlap.
To avoid this, this change introduce a minimum distance from the last label that
has to be respected for the next label to be draw too, otherwise it is skipped.
Only the first and last labels in any given direction are drawn every time.

The distance was chosen to avoid overlap between 2 consecutive 2-digits labels.

Before: https://i.imgur.com/2RHqi5M.jpg
After: https://i.imgur.com/AgHwPSZ.jpg

Bug: 1098673
Change-Id: Ib5ae09ef8c212707d45e7c0b602599c3a8011454