Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> What's the latest here? Is this still going ahead?

Yes, sorry about the lack of updates.

From the feedback [1], it sounded like we will need to include some kind of attribution to the dropped frames before this can go ahead. This is going to need some additional work (related bug [2]), and we are waiting on that for now (we are not blocked, but also don't have the resources right now). We do hope to find resources to make this happen in ~Q3.

The UX changes suggested by Paul [3] is also going to need some changes, but hopefully that will not be too much work.

Does that make sense? Or would it make sense to start without the attribution in the first cut, and add attribution when it is available?

[1] https://docs.google.com/document/d/1-KP3fAjemdm7lnvCll9T1Lg-ikCzRPI-oNCp8fHAQCc/edit?disco=AAAAGdS88VQ
[2] https://bugs.chromium.org/p/chromium/issues/detail?id=1086974
[3] https://docs.google.com/document/d/1-KP3fAjemdm7lnvCll9T1Lg-ikCzRPI-oNCp8fHAQCc/edit?disco=AAAAJepCbR0

Patch-set: 5